Tema

+++ src/main/java/com/generation/blogpessoal/controller/PostagemController.java
@@ -1,6 +1,7 @@
 package com.generation.blogpessoal.controller;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -15,9 +16,11 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 import com.generation.blogpessoal.model.Postagem;
 import com.generation.blogpessoal.repository.PostagemRepository;
+import com.generation.blogpessoal.repository.TemaRepository;
 
 import jakarta.validation.Valid;
 
@@ -28,6 +31,9 @@ public class PostagemController {
 
 	@Autowired
 	private PostagemRepository postagemRepository;
+	
+	@Autowired
+	private TemaRepository temaRepository; // ajustar o metodo put e push//
 
 	@GetMapping
 	public ResponseEntity<List<Postagem>> getAll() {
@@ -63,21 +69,28 @@ public class PostagemController {
 	 
 	}
 	
-	@PutMapping	
+	@PutMapping
 	public ResponseEntity<Postagem> put(@Valid @RequestBody Postagem postagem){
-			return ResponseEntity.status(HttpStatus.OK)
-				.body(postagemRepository.save(postagem));
-	
-				/* UPDATE tb_postagens SET titulo = ?, texto = ?, data = ?
-				 * *WHERE id = id*/
-				 
+		
+		
+		return postagemRepository.findById(postagem.getId())
+				.map(resposta -> ResponseEntity.status(HttpStatus.OK).body(postagemRepository.save(postagem)))
+				.orElse(ResponseEntity.status(HttpStatus.NOT_FOUND).build());
+		// UPDATE tb_postagens SET titulo = "???", texto = "???", data = "???" WHERE id = ?
 	}
 	
 	@ResponseStatus(HttpStatus.NO_CONTENT)
-	@DeleteMapping("/{id}")
-	public void delete (@PathVariable Long id) {
-		postagemRepository.deleteById(id);
+	@DeleteMapping("{id}")
+	public void delete(@PathVariable Long id) {
+		// DELETE FROM tb_postagens WHERE id = ?
 		
-		/* DELETE FROM tb_postagens WHERE id = ?*/
+		Optional<Postagem> optional = postagemRepository.findById(id);
+		
+		if (optional.isEmpty()) {
+			throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+		}
+		else {
+			postagemRepository.deleteById(id);
+		}
 	}
 }
